tempate ext_proc client
