+++ source/extensions/filters/common/ext_proc/client_base.h
@@ -7,23 +7,25 @@
 namespace Envoy {
 namespace Extensions {
 namespace Common {
-namespace ExtProc {
+namespace ExternalProcessing {
 
 // Forward declarations for templated classes
-template <typename RequestType, typename ResponseType>
-class RequestCallbacks;
+template <typename ResponseType> class RequestCallbacks;
 
-template <typename RequestType, typename ResponseType>
-class StreamBase;
+template <typename RequestType, typename ResponseType> class ClientBase;
 
-template <typename RequestType, typename ResponseType>
-class ClientBase;
+/**
+ * Stream base class used during external processing.
+ */
+class StreamBase {
+public:
+  virtual ~StreamBase() = default;
+};
 
 /**
  * Async callbacks used during external processing.
  */
-template <typename RequestType, typename ResponseType>
-class RequestCallbacks {
+template <typename ResponseType> class RequestCallbacks {
 public:
   virtual ~RequestCallbacks() = default;
   virtual void onComplete(ResponseType& response) PURE;
@@ -31,24 +33,13 @@ public:
 };
 
 /**
- * Stream base class used during external processing.
- */
-template <typename RequestType, typename ResponseType>
-class StreamBase {
-public:
-  virtual ~StreamBase() = default;
-};
-
-/**
  * Async client base class used during external processing.
  */
-template <typename RequestType, typename ResponseType>
-class ClientBase {
+template <typename RequestType, typename ResponseType> class ClientBase {
 public:
   virtual ~ClientBase() = default;
   virtual void sendRequest(RequestType&& request, bool end_stream, const uint64_t stream_id,
-                           RequestCallbacks<RequestType, ResponseType>* callbacks,
-                           StreamBase<RequestType, ResponseType>* stream) PURE;
+                           RequestCallbacks<ResponseType>* callbacks, StreamBase* stream) PURE;
   virtual void cancel() PURE;
   virtual const Envoy::StreamInfo::StreamInfo* getStreamInfo() const PURE;
 };
@@ -56,7 +47,7 @@ public:
 template <typename RequestType, typename ResponseType>
 using ClientBasePtr = std::unique_ptr<ClientBase<RequestType, ResponseType>>;
 
-} // namespace ExtProc
+} // namespace ExternalProcessing
 } // namespace Common
 } // namespace Extensions
 } // namespace Envoy
