+++ source/extensions/filters/http/ext_proc/config.cc
@@ -94,8 +94,8 @@ ExternalProcessingFilterConfig::createFilterFactoryFromProtoTyped(
   if (proto_config.has_grpc_service()) {
     return [filter_config = std::move(filter_config), &context,
             dual_info](Http::FilterChainFactoryCallbacks& callbacks) {
-      auto client = std::make_unique<ExternalProcessorClientImpl>(
-          context.clusterManager().grpcAsyncClientManager(), dual_info.scope);
+      auto client = createExternalProcessorClient(context.clusterManager().grpcAsyncClientManager(),
+                                                  dual_info.scope);
       callbacks.addStreamFilter(
           Http::StreamFilterSharedPtr{std::make_shared<Filter>(filter_config, std::move(client))});
     };
@@ -139,7 +139,7 @@ ExternalProcessingFilterConfig::createFilterFactoryFromProtoWithServerContextTyp
   if (proto_config.has_grpc_service()) {
     return [filter_config = std::move(filter_config),
             &server_context](Http::FilterChainFactoryCallbacks& callbacks) {
-      auto client = std::make_unique<ExternalProcessorClientImpl>(
+      auto client = createExternalProcessorClient(
           server_context.clusterManager().grpcAsyncClientManager(), server_context.scope());
       callbacks.addStreamFilter(
           Http::StreamFilterSharedPtr{std::make_shared<Filter>(filter_config, std::move(client))});
