+++ source/extensions/filters/http/ext_proc/http_client/BUILD
@@ -17,7 +17,7 @@ envoy_cc_library(
         "//source/common/common:enum_to_int",
         "//source/common/http:header_map_lib",
         "//source/common/http:utility_lib",
-        "//source/extensions/filters/http/ext_proc:client_base_interface",
+        "//source/extensions/filters/http/ext_proc:client_lib",
         "@envoy_api//envoy/extensions/filters/http/ext_proc/v3:pkg_cc_proto",
         "@envoy_api//envoy/service/ext_proc/v3:pkg_cc_proto",
     ],
