+++ source/extensions/filters/http/ext_proc/http_client/http_client_impl.h
@@ -7,14 +7,18 @@
 #include "envoy/service/ext_proc/v3/external_processor.pb.h"
 
 #include "source/common/common/logger.h"
-#include "source/extensions/filters/http/ext_proc/client_base.h"
+#include "source/extensions/filters/common/ext_proc/client_base.h"
+#include "source/extensions/filters/http/ext_proc/client_impl.h"
 
 namespace Envoy {
 namespace Extensions {
 namespace HttpFilters {
 namespace ExternalProcessing {
 
-class ExtProcHttpClient : public ClientBase,
+// Now use the templated base class with the HTTP-specific types
+class ExtProcHttpClient : public Envoy::Extensions::Common::ExternalProcessing::ClientBase<
+                              envoy::service::ext_proc::v3::ProcessingRequest,
+                              envoy::service::ext_proc::v3::ProcessingResponse>,
                           public Http::AsyncClient::Callbacks,
                           public Logger::Loggable<Logger::Id::init> {
 public:
@@ -25,8 +29,10 @@ public:
   ~ExtProcHttpClient() override { cancel(); }
 
   void sendRequest(envoy::service::ext_proc::v3::ProcessingRequest&& req, bool end_stream,
-                   const uint64_t stream_id, RequestCallbacks* callbacks,
-                   StreamBase* stream) override;
+                   const uint64_t stream_id,
+                   Envoy::Extensions::Common::ExternalProcessing::RequestCallbacks<
+                       envoy::service::ext_proc::v3::ProcessingResponse>* callbacks,
+                   Envoy::Extensions::Common::ExternalProcessing::StreamBase* stream) override;
   void cancel() override;
   void onBeforeFinalizeUpstreamSpan(Tracing::Span&, const Http::ResponseHeaderMap*) override {}
 
@@ -45,7 +51,8 @@ private:
   envoy::extensions::filters::http::ext_proc::v3::ExternalProcessor config_;
   Server::Configuration::ServerFactoryContext& context_;
   Http::AsyncClient::OngoingRequest* active_request_{};
-  RequestCallbacks* callbacks_{};
+  Envoy::Extensions::Common::ExternalProcessing::RequestCallbacks<
+      envoy::service::ext_proc::v3::ProcessingResponse>* callbacks_{};
 };
 
 } // namespace ExternalProcessing
